Change all imports to material-ui/Tabs
