+++ newIDE/app/src/UI/Tabs.js
@@ -1,174 +1,40 @@
-import React, { Component } from 'react';
-import { Tabs as MaterialUITabs, Tab as MaterialUITab } from 'material-ui/Tabs';
-import Close from 'material-ui/svg-icons/navigation/close';
-import FlatButton from 'material-ui/FlatButton';
-import ThemeConsumer from './Theme/ThemeConsumer';
+// @flow
+import * as React from 'react';
+import { Tabs as MUITabs, Tab as MUITab } from 'material-ui/Tabs';
 
-const styles = {
-  tabsContainerStyle: {
-    maxWidth: '100%',
-    flex: 1,
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  tabTemplateStyle: {
-    width: '100%',
-    position: 'relative',
-    textAlign: 'initial',
-    height: '100%',
-  },
-};
+// We support a subset of the props supported by Material-UI v0.x Tabs
+// They should be self descriptive - refer to Material UI docs otherwise.
+type TabsProps<TabName> = {|
+  value?: TabName,
+  onChange: TabName => void,
+  children: React.Node, // Should be Tab
+|};
 
 /**
- * This is to override the default material-ui tab template used to wrap the content of each tab element.
- * 2 changes (**please port them to your new implementation if you change the tabs**):
- *
- * 1) Instead of setting the "height" of hidden tabs to "0", we set "display" to "none" to avoid
- * messing with components (in particular components where you can scroll: when collapsed because of height=0,
- * they will lose they scrolling position).
- *
- * 2) shouldComponentUpdate is used to avoid updating the content of a tab that is not selected.
- *
- * Rest of the implementation is the same.
+ * Tabs based on Material-UI Tabs.
  */
-class TabTemplate extends Component {
-  shouldComponentUpdate(nextProps) {
-    return this.props.selected || nextProps.selected;
-  }
-
+export class Tabs<TabName> extends React.Component<TabsProps<TabName>, {||}> {
   render() {
-    const { children, selected, style } = this.props;
-    const templateStyle = { ...styles.tabTemplateStyle, ...style };
-    if (!selected) {
-      templateStyle.display = 'none';
-    }
-
-    return <div style={templateStyle}>{children}</div>;
+    return <MUITabs {...this.props} />;
   }
 }
 
-export class Tabs extends Component {
-  render() {
-    const { hideLabels, ...tabsProps } = this.props;
+// We support a subset of the props supported by Material-UI v0.x Table
+// They should be self descriptive - refer to Material UI docs otherwise.
+type TabProps = {|
+  label: React.Node,
+  value: string,
+  children?: React.Node,
+|};
 
-    return (
-      <ThemeConsumer>
-        {muiTheme => {
-          const tabItemContainerStyle = {
-            maxWidth: '100%', // Tabs should take all width
-            flexShrink: 0, // Tabs height should never be reduced
-            overflowX: 'auto',
-            display: hideLabels ? 'none' : 'block',
-            backgroundColor: muiTheme.closableTabs.backgroundColor,
-          };
-
-          const contentContainerStyle = {
-            overflowY: 'hidden',
-            height: hideLabels
-              ? '100%'
-              : `calc(100% - ${muiTheme.closableTabs.height}px)`,
-          };
-
-          return (
-            <MaterialUITabs
-              style={styles.tabsContainerStyle}
-              tabTemplate={TabTemplate}
-              contentContainerStyle={contentContainerStyle}
-              tabItemContainerStyle={tabItemContainerStyle}
-              inkBarStyle={{ display: 'none' }}
-              {...tabsProps}
-            />
-          );
-        }}
-      </ThemeConsumer>
-    );
-  }
-}
-Tabs.muiName = 'Tabs';
-
-export class Tab extends Component {
+/**
+ * A Tab based on Material-UI Tab.
+ */
+export class Tab extends React.Component<TabProps, {||}> {
+  // Set muiName to let Material-UI's v0.x Table recognise
+  // the component.
   static muiName = 'Tab';
-
   render() {
-    const { selected, onClose, label, closable, ...tabProps } = this.props;
-
-    return (
-      <ThemeConsumer>
-        {muiTheme => {
-          const truncatedLabel = (
-            <span
-              style={{
-                width:
-                  muiTheme.closableTabs.width -
-                  muiTheme.closableTabs.closeButtonWidth * 1.5,
-                marginRight: muiTheme.closableTabs.closeButtonWidth,
-                overflow: 'hidden',
-                textOverflow: 'ellipsis',
-              }}
-            >
-              {label}
-            </span>
-          );
-
-          return (
-            <span
-              style={{
-                position: 'relative',
-                width: muiTheme.closableTabs.width,
-                display: 'inline-block',
-              }}
-            >
-              <MaterialUITab
-                {...tabProps}
-                label={truncatedLabel}
-                selected={selected}
-                buttonStyle={{
-                  height: muiTheme.closableTabs.height,
-                  backgroundColor: selected
-                    ? muiTheme.closableTabs.selectedBackgroundColor
-                    : muiTheme.closableTabs.backgroundColor,
-                  color: selected
-                    ? muiTheme.closableTabs.selectedTextColor
-                    : muiTheme.closableTabs.textColor,
-                }}
-                style={{
-                  height: '100%',
-                  width: '100%',
-                }}
-              />
-              {closable && (
-                <FlatButton
-                  backgroundColor="transparent"
-                  hoverColor={muiTheme.selectedBackgroundColor}
-                  onClick={onClose}
-                  style={{
-                    position: 'absolute',
-                    right: 0,
-                    top: 0,
-                    bottom: 0,
-                    borderRadius: 0,
-                    width: muiTheme.closableTabs.closeButtonWidth,
-                    minWidth: muiTheme.closableTabs.closeButtonWidth,
-                  }}
-                  icon={
-                    <Close
-                      color={
-                        selected
-                          ? muiTheme.closableTabs.selectedTextColor
-                          : muiTheme.closableTabs.textColor
-                      }
-                      style={{
-                        width: muiTheme.closableTabs.height / 2,
-                        height: muiTheme.closableTabs.height / 2,
-                      }}
-                    />
-                  }
-                />
-              )}
-            </span>
-          );
-        }}
-      </ThemeConsumer>
-    );
+    return <MUITab {...this.props} />;
   }
 }
